requirements.txt
bcrypt==4.0.1
cryptography
fastapi[standard]~=0.115.5
psycopg2-binary
pydantic~=2.10.1
PyMySQL
python-jose~=3.3.0
python-multipart
SQLAlchemy~=2.0.36
uvicorn[standard]
passlib~=1.7.4
pytest
httpx
pytest-asyncio
aiofiles
jinja2
starlette~=0.41.3
alembic~=1.14.0